+++ langgraph/backend/app/main.py
@@ -5,9 +5,6 @@ from app.graphs.graph import build_main_app
 from app.nodes.supervisor import make_node_supervisor
 from app.nodes.problem_framing import make_node_problem_framing
 from app.agents.problem_framing.agent import build_agent as build_pf_agent
-from app.agents.translator.agent import build_agent as build_translation_agent
-from app.nodes.detect_language import make_node_detect_language
-from app.nodes.translation import make_node_translation
 
 from app.ui.ui import SageCompassUI
 
@@ -22,21 +19,14 @@ def build_app_for_cli():
 
     node_supervisor = make_node_supervisor()
 
-    translation_agent = build_translation_agent()
-    translation_node = make_node_translation(translation_agent)
-
-    node_detect_language = make_node_detect_language()
-
     return build_main_app(
-        detect_language_node=node_detect_language,
         supervisor_node=node_supervisor,
         problem_framing_node=node_problem_framing,
-        translation_node=translation_node,
     )
 
 
 def main():
-    maybe_attach_pycharm()
+    # maybe_attach_pycharm()
     app = build_app_for_cli()
     ui = SageCompassUI(app)
     ui.launch()